pyproject.toml
[tool.poetry]
name = "async-python-sprint-4"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "0.85.0"
uvicorn = {extras = ["standart"], version = "0.18.3"}
pydantic = {extras = ["email", "dotenv"], version = "1.9.1"}

psycopg2-binary = "2.9.3"
sqlalchemy = {version = "1.4.41", extras = ["asyncio"]}
sqlalchemy-utils = "0.38.3"
alembic = "1.8.1"
asyncpg = "0.26.0"
greenlet = "1.1.2"

JSON-log-formatter = "0.3.1"
python-json-logger = "2.0.1"
python-dateutil = "2.8.2"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-multipart = "^0.0.5"
validate_email = "^1.3"
pytest = "^7.2.0"
httpx = "^0.23.1"
requests = "^2.28.1"
aiocache =  {extras = ["redis"], version = "0.11.1"}
aioredis = "^2.0.1"

[tool.poetry.dev-dependencies]
pylama = {extras = ["toml", "pylint", "pycodestyle", "pyflakes"], version = "8.4.1"}
isort = "5.10.1"
black = "22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylama]
linters = "pylint,pycodestyle,pyflakes"
max_line_length = 79
skip = "**/migrations/*,**/migration_tool/*"
format = "pylint"
ignore = "C0114,C0115,C0116,C0103,C0121,C0411,R0903,W0107,W1203,W0511,E0611,E0213,E0401"

[[tool.pylama.files]]
path = "**/__init__.py"
ignore = "W0611"

[tool.black]
line-length = 79
target-version = ['py310']
extend-exclude = "/migrations|migration_tool/"

[tool.isort]
py_version = 310
profile = "black"
line_length = 79
skip_glob = ["**/migrations/*", "**/migration_tool/*"]
known_local_folder = [
    "api",
    "db",
    "dto",
    "services",
    "settings",
    "tests",
    "utils",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
